+++ app/src/renderer/components/staking/UndelegationModal.vue
@@ -35,11 +35,31 @@
       field-label="To"
     >
       <tm-field id="to" v-model="to" readonly="readonly" />
+      <hr />
+    </tm-form-group>
+    <tm-form-group
+      class="undelegation-modal-form-group"
+      field-id="password"
+      field-label="Account password"
+    >
+      <tm-field
+        id="password"
+        v-model="password"
+        :type="showPassword ? `text` : `password`"
+        placeholder="password..."
+      />
+      <input
+        id="showPasswordCheckbox"
+        v-model="showPassword"
+        type="checkbox"
+        @input="togglePassword"
+      />
+      <label for="showPasswordCheckbox">Show password</label>
     </tm-form-group>
     <div class="undelegation-modal-footer">
       <tm-btn
         id="submit-undelegation"
-        :disabled="$v.amount.$invalid"
+        :disabled="$v.$invalid"
         color="primary"
         value="Undelegate"
         size="lg"
@@ -81,7 +101,9 @@ export default {
     }
   },
   data: () => ({
-    amount: 0
+    amount: 0,
+    password: ``,
+    showPassword: false
   }),
   computed: {
     ...mapGetters([`bondingDenom`])
@@ -92,6 +114,9 @@ export default {
         required,
         isInteger,
         between: between(1, this.maximum)
+      },
+      password: {
+        required
       }
     }
   },
@@ -99,9 +124,13 @@ export default {
     close() {
       this.$emit(`update:showUndelegationModal`, false)
     },
+    togglePassword() {
+      this.showPassword = !this.showPassword
+    },
     onUndelegate() {
       this.$emit(`submitUndelegation`, {
-        amount: this.amount
+        amount: this.amount,
+        password: this.password
       })
       this.close()
     }